Add command line parser
