+++ core/src/main/scala/zio/config/package.scala
@@ -22,6 +22,9 @@ package object config extends ReadFunctions with WriteFunctions with ConfigDocsF
   private[config] def seqEither[A, B](either: List[Either[A, B]]): Either[A, List[B]] =
     either.foldRight(Right(List.empty[B]): Either[A, List[B]])((a, b) => a.flatMap(aa => b.map(bb => aa :: bb)))
 
+  private[config] def seqOption[A](options: List[Option[A]]): Option[List[A]] =
+    options.foldRight(Some(Nil): Option[List[A]])((a, b) => a.flatMap(aa => b.map(bb => aa :: bb)))
+
   private[config] def seqEither2[A, B, C](genError: (Int, A) => C)(list: List[Either[A, B]]): Either[List[C], List[B]] =
     list.zipWithIndex
       .foldLeft(Right(Nil): Either[List[C], List[B]]) {
